refactor: split DTOs of api
